+++ dynawo/sources/Models/CPP/ModelNetwork/test/TestTwoWindingsTransformer.cpp
@@ -179,7 +179,7 @@ createModelTwoWindingsTransformer(bool open, bool initModel, bool ratioTapChange
     double* f1 = new double[bus1->sizeF()];
     double* z1 = new double[bus1->sizeZ()];
     bool* zConnected1 = new bool[bus1->sizeZ()];
-    for (size_t i = 0; i < bus1->sizeZ(); ++i)
+    for (int i = 0; i < bus1->sizeZ(); ++i)
       zConnected1[i] = true;
     bus1->setReferenceZ(&z1[0], zConnected1, 0);
     bus1->setReferenceY(y1, yp1, f1, 0, 0);
@@ -201,7 +201,7 @@ createModelTwoWindingsTransformer(bool open, bool initModel, bool ratioTapChange
     double* f2 = new double[bus2->sizeF()];
     double* z2 = new double[bus2->sizeZ()];
     bool* zConnected2 = new bool[bus2->sizeZ()];
-    for (size_t i = 0; i < bus2->sizeZ(); ++i)
+    for (int i = 0; i < bus2->sizeZ(); ++i)
       zConnected2[i] = true;
     bus2->setReferenceZ(&z2[0], zConnected2, 0);
     bus2->setReferenceY(y2, yp2, f2, 0, 0);
@@ -256,7 +256,7 @@ TEST(ModelsModelNetwork, ModelNetworkTwoWindingsTransformerCalculatedVariables) 
   std::vector<double> f(t2w->sizeF(), 0.);
   std::vector<double> z(t2w->sizeZ(), 0.);
   bool* zConnected = new bool[t2w->sizeZ()];
-  for (size_t i = 0; i < t2w->sizeZ(); ++i)
+  for (int i = 0; i < t2w->sizeZ(); ++i)
     zConnected[i] = true;
   t2w->setReferenceZ(&z[0], zConnected, 0);
   t2w->setReferenceY(&y[0], &yp[0], &f[0], 0, 0);
@@ -475,7 +475,7 @@ TEST(ModelsModelNetwork, ModelNetworkTwoWindingsTransformerCalculatedVariablesOp
   std::vector<double> f(t2w->sizeF(), 0.);
   std::vector<double> z(t2w->sizeZ(), 0.);
   bool* zConnected = new bool[t2w->sizeZ()];
-  for (size_t i = 0; i < t2w->sizeZ(); ++i)
+  for (int i = 0; i < t2w->sizeZ(); ++i)
     zConnected[i] = true;
   t2w->setReferenceZ(&z[0], zConnected, 0);
   t2w->setReferenceY(&y[0], &yp[0], &f[0], 0, 0);
